R-type works
From what I can tell R-type works.

changed address size to 10 bits in dpram, pc_incr, program_counter

changed flagsReg to have an enable because if it dosen't have it then the flags are changed before the value is saved in the regFile. This is fine until we do branch and jump where we need the flags.

rdest_decoder added enable because if it doesn't have it the value is saved into the regFile multiple times.

Alot was changed in the CPU but comments talk though most of it